High Scores and Screens
Added high score saving between sessions and two
basic screens, a title screen that shows when the
game first boots up and a game over screen that shows
when the player dies (and can alert them of a new
high score).

Also a major change was that I changed the core
node from a Label to a simple Node, and instead
instantiated the label and all it's font options
dynamically from script.